requirements.txt
django
django-crispy-forms
Pillow
numpy
matplotlib
wfastcgi
twilio